UP: Add a user process starting framework.
This patch adds the 'user_process' macro that, when expanded, will
create a list of structs in the .user_processes section of the ELF
file.

We also define two variables in the linker script that point to the
beginning and end of this list so it can be easily traversed.

Signed-off-by: Matthew Leach <matthew@mattleach.net>