io: fix low-level IO functions
When using the io functions in code, we made assumptions to do with
which registers the IO values and addresses where in. This is fine if
we use these as function calls but since we're inlining these
functions, we need to tell the which C expressions these values
correspond to.

Signed-off-by: Matthew Leach <matthew@mattleach.net>