code clean: Remove dead code
Since POS is now targeting the vexpress-a9 machine, we can get rid of
legacy drivers and syscalls that we no longer need.

Because we are removing syscalls, we need to change the svc numbers of
the ones we are leaving behind. Also, remove user processes that made
use of the syscalls that we are removing and replace it with a simple
"hello world from userspace" program.

Signed-off-by: Matthew Leach <matthew@mattleach.net>