mm: Add kernel virtual memory.
The kernel currently is linked at an address that is within the
physical RAM of the platform. Ideally we would like the kernel to be
linked at a virtual address and use page tables so we don't have to
worry about changing the link address.

This patch enables virtual memory on an ARMv7 platform and plants the
kernel at a virtual "low memory" address.

Signed-off-by: Matthew Leach <matthew@mattleach.net>